requirements-travis.txt
stem
pytest-cov
psutil
